Generalised the image tagging so the functionality can be reused in the new commands.
